ref(spans): Graduate standalone span ingestion
