+++ relay-server/src/processing/spans/filter.rs
@@ -7,14 +7,6 @@ use crate::managed::Managed;
 use crate::processing::Context;
 use crate::processing::spans::{Error, ExpandedSpans, Result, SerializedSpans};
 
-/// Filters standalone spans sent for a project which does not allow standalone span ingestion.
-pub fn feature_flag(ctx: Context<'_>) -> Result<()> {
-    match ctx.should_filter(Feature::StandaloneSpanIngestion) {
-        true => Err(Error::FilterFeatureFlag),
-        false => Ok(()),
-    }
-}
-
 // Filters span attachments for a project which does not allow for span attachment ingestion.
 pub fn feature_flag_attachment(
     spans: Managed<SerializedSpans>,
